Extract formatters for Samples, Cache, Package; move PackageViews to Views/
- Create SamplesOutputFormatter with FormatSamplesList, WriteSamplesWithContent, WriteSamplesTitle
- Create CacheOutputFormatter with FormatCacheInfo, FormatSize
- Create PackageOutputFormatter with WriteFileTree (file tree rendering)
- Move FileTreeView, PackageDependenciesView, and their contexts to Views/PackageViews.cs
- Remove ad-hoc Console.WriteLine formatting from CacheCommand
- Remove inline Markout rendering from SamplesCommand and PackageCommand
- All 253 tests pass
